integrate getDocument with existing service
