+++ frontend/fingertips-frontend/lib/search/indicatorSearchService.ts
@@ -57,4 +57,8 @@ export class IndicatorSearchService implements IIndicatorSearchService {
 
     return results.slice(0, 20);
   }
+
+  async getIndicator(indicatorId: string): Promise<IndicatorDocument> {
+    return this.searchClient.getDocument(indicatorId);
+  }
 }
