+++ frontend/fingertips-frontend/lib/search/searchTypes.ts
@@ -30,12 +30,9 @@ export interface IIndicatorSearchService {
     searchTerm: string,
     areaCodes?: string[]
   ): Promise<IndicatorDocument[]>;
+  getIndicator(indicatorId: string): Promise<IndicatorDocument>;
 }
 
 export interface IAreaSearchService {
   getAreaSuggestions(partialAreaName: string): Promise<AreaDocument[]>;
-}
-
-export interface IIndicatorDocumentService {
-  getIndicator(indicatorId: string): Promise<IndicatorDocument>;
 }
